NEW: Add Aircraft Carrier Water Demon, build from Hadal Hydrothermal Vortex. NEW: Add repeat build mode for Abyssal Hydrothermal Vent and Hadal Hydrothermal Vortex. (click build button twice) NOTE: For Hadal Hydrothermal Vortex, the build materials will set automatically.       For Abyssal Hydrothermal Vent, record the materials when clicking the button to loop mode. NEW: Support FluidContainerRegistry and IFluidContainerItem for fuel (lava only). NOTE: For IFluidContainerItem (ex: ender tank, universal cell), the lava will not extract when stack size > 1. NEW: Add recipe: Grudge block + Iron block -> Abyssium Block NEW: Support forge ore dictionary: Polymetallic Ore = "oreManganese", Polymetallic Nodules = "dustManganese". FIX: Fix the emotion changes too fast bug (< 5 sec) when sitting. FIX: Fix some particle bug. CHANGE: Tweak mount and ship idle AI. CHANGE: Tweak some equip value. CHANGE: Tweak models and textures. CHANGE: Tweak bounding box and eye height for Morph mod. CHANGE: Hostile ship will spawn in the water block of BEACH and WATER(OCEAN, RIVER) biome. CHANGE: Hide unavailable AI options in ship inventory GUI. CHANGE: Hadal Hydrothermal Vortex now accept polymetal gravel. CHANGE: Cancel the damage reducing and armor(%) if attacker is owner. CHANGE: (friendly ship only) Change the damage reducing of ship vs ship according to the armor(%) in config: "ship_limit". CHANGE: Dismantle Hammer changes: left click: one shot kill (owner only), sneaking + right click: dismantle.

+++ src/main/java/com/lulan/shincolle/client/inventory/ContainerSmallShipyard.java
@@ -69,7 +69,7 @@ public class ContainerSmallShipyard extends Container {
 	public ItemStack transferStackInSlot(EntityPlayer player, int slotid) {
         ItemStack itemstack = null;
         Slot slot = (Slot)this.inventorySlots.get(slotid);
-        byte itemID = -1;
+        int itemID = -1;
 
         if(slot != null && slot.getHasStack()) { 			//若slot有東西
             ItemStack itemstack1 = slot.getStack();			//itemstack1取得該slot物品
